backend/requirements.txt
flask
flask-smorest
flask-cors
marshmallow
python-dotenv
